requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-base-url==2019.7.24
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-jwt==1.8.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
markdown2==2.3.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.2
Pillow==6.2.0
PyJWT==1.4.0
pyparsing==2.4.2
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
